Move emby to libraries
