+++ libraries/emby/client.py
