+++ libraries/emby/core/connection_manager.py
