+++ libraries/emby/core/credentials.py
