+++ libraries/emby/core/exceptions.py
